+++ pwndbg/commands/kmem_trace.py
